requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
docopt==0.6.2
Flask==2.2.2
frozenlist==1.3.3
future==0.18.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
line-bot-sdk==2.4.1
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.8
pipreqs==0.4.11
pycodestyle==2.10.0
python-dotenv==0.21.1
requests==2.28.1
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.14
Werkzeug==2.2.2
yarg==0.1.9
yarl==1.8.2
